Rename base solvers tests class
