+++ tests/mesolve/test_rouchon.py
@@ -4,7 +4,7 @@ import pytest
 
 import dynamiqs as dq
 
-from .mesolver_tester import MEAdjointSolverTester
+from .me_solver_tester import MEAdjointSolverTester
 from .open_system import LeakyCavity
 
 leaky_cavity_8 = LeakyCavity(n=8, kappa=2 * pi, delta=2 * pi, alpha0=1.0)
