+++ tests/sesolve/se_solver_tester.py
